use half3

+++ zenovis/xinxinoptix/optixPathTracer.h
@@ -160,12 +160,12 @@ struct Params
     float4*      accum_buffer_T;
     float4*      accum_buffer_B;
     uchar4*      frame_buffer;
-    float3*      frame_buffer_C;
-    float3*      frame_buffer_D;
-    float3*      frame_buffer_S;
-    float3*      frame_buffer_T;
-    float3*      frame_buffer_B;
-    float3*      frame_buffer_M;
+    ushort3*     frame_buffer_C;
+    ushort3*     frame_buffer_D;
+    ushort3*     frame_buffer_S;
+    ushort3*     frame_buffer_T;
+    ushort3*     frame_buffer_B;
+    ushort3*     frame_buffer_M;
 
     float3*      debug_buffer;
     float3*      albedo_buffer;
